+++ tests/testthat/test-vetted_models-survreg.R
@@ -167,14 +167,4 @@ test_that("vetted_models survreg()", {
       add_q(),
     NA
   )
-  expect_warning(
-    trial %>%
-      tbl_uvregression(
-        y = Surv(ttdeath, death),
-        method = survreg
-      ) %>%
-      add_global_p() %>%
-      add_q(),
-    NA
-  )
 })